Inconsistent use of package clockface and clockface_test
There is inconsistent use of package names in these examples. 

For example, the packages clockface and clockface_test are used early one. To make this work, the clockface_test.go file imports clockface, but this changes in later files and is really unclear to a beginner how to fix. 

I thought I'd screwed up and cludged my way through it, but on third look I think it would be much simpler to fix this to use a single package and have examples that work directly if transposed or copied. 

Fair warning - I am a novice golang coder so if I am missing some larger picture around the change from 1 package to 2 please let me know!